+++ lib/page/staking/index.dart
@@ -7,7 +7,6 @@ import 'package:auro_wallet/page/staking/components/delegationInfo.dart';
 import 'package:auro_wallet/page/staking/components/stakingOverview.dart';
 import 'package:auro_wallet/service/api/api.dart';
 import 'package:auro_wallet/store/app.dart';
-import 'package:auro_wallet/utils/UI.dart';
 import 'package:flutter/material.dart';
 
 class Staking extends StatefulWidget {
@@ -26,6 +25,8 @@ class _StakingState extends State<Staking> {
   bool loading = true;
   Timer? _refreshTimer;
   ReactionDisposer? _storeChangeDisposer;
+  final GlobalKey<RefreshIndicatorState> _stakingRefreshKey =
+      GlobalKey<RefreshIndicatorState>();
 
   @override
   void initState() {
@@ -34,6 +35,7 @@ class _StakingState extends State<Staking> {
     final hasCachedData = store.assets!.mainTokenNetInfo.tokenBaseInfo != null;
     loading = store.staking!.lastLoadedDataKey != currentKey || !hasCachedData;
     WidgetsBinding.instance.addPostFrameCallback((_) {
+      store.setStakingRefreshKey(_stakingRefreshKey);
       _fetchData();
       _refreshTimer = Timer.periodic(Duration(minutes: 3), (timer) {
         _onRefresh();
@@ -74,6 +76,7 @@ class _StakingState extends State<Staking> {
   void dispose() {
     _refreshTimer?.cancel();
     _storeChangeDisposer?.call();
+    store.setStakingRefreshKey(null);
     super.dispose();
   }
 
@@ -99,7 +102,7 @@ class _StakingState extends State<Staking> {
         child: RefreshIndicator(
             backgroundColor: Colors.white,
             color: Theme.of(context).primaryColor,
-            key: globalStakingRefreshKey,
+            key: _stakingRefreshKey,
             onRefresh: _fetchData,
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,