+++ lib/utils/UI.dart
@@ -573,8 +573,4 @@ class UI {
   }
 }
 
-final GlobalKey<RefreshIndicatorState> globalStakingRefreshKey =
-    new GlobalKey<RefreshIndicatorState>();
 
-final GlobalKey<RefreshIndicatorState> globalTokenRefreshKey =
-    new GlobalKey<RefreshIndicatorState>();